Created an IgnoreCollision script, MirrorMove script, and a PlaySoundOnCollision script. Updated the PlayerMove script as well
The player move script was updated to settle on rounded numbers after you finish
moving.

The ignore collision script will let an object ignore another objects
collision but you need to to add it to both if you want them both to ignore
collision.

The MirrorMove script will mirror the movement of another game object. This also
settles the object on rounded numbers.

The PlaySoundOnCollision script will play the sound on collision. It needs an
audio source on the object though too.